requirements.txt
nonebot[scheduler]
aiocqhttp
lxml>=4.4.1
pytz
requests
sogou_tr_free>=0.0.6
zhconv
Pillow
matplotlib
numpy>=1.18.0
beautifulsoup4>=4.9.0
pygtrie
tinydb>=4.0
aiohttp
peony-twitter[all]==2.1.2

ujson
httpx==0.22.0
peewee
sqlitedict==1.7.0
PyYAML~=6.0
Quart==0.16.3
pandas
xlwt
qrcode
anyio